CLIP-DINOV2-Image-Similarity/requirements.txt
transformers
Pillow
torch